feat: user able to unsubscribe from Telegram

+++ infrastructure/ISubscriptionRepository.cs
@@ -6,6 +6,6 @@ public interface ISubscriptionRepository
 {
     Task<List<Subscription>> GetAll();
     Task Add(long userId, string board, string keyword);
-    Task Delete(int userId, string board, string keyword);
+    Task Delete(long userId, string board, string keyword);
     Task<List<Subscription>> Get(string board);
 }
